Fixed remaining pep8 Python2 compatibility
